refactor: lean-component 규칙 적용 — 컴포넌트 파일 경량화
새 파일 (분리된 타입·상수·유틸·서브컴포넌트):
- src/types/board.ts: BoardEmptyProps, CourseRowProps, DayChipProps, MenuBoardProps, PreMealPickProps
- src/types/shared.ts: ErrorBoundaryProps, ErrorBoundaryState
- src/types/site.ts: EntryInnerProps
- src/constants/board.ts: BOARD_EMPTY_COPY
- src/constants/date.ts: DOW
- src/lib/hero-status.ts: HeroStatusState, nextWorkdayKey, getHeroStatus
- src/lib/day-bar.ts: chipBg, dowColor, labelClass, dateClass
- src/components/@shared/ErrorFallback.tsx
- src/components/board/DayChip.tsx
- src/components/home/EntryInner.tsx
- src/components/home/HomeSide.styles.ts

기존 파일 정리:
- src/constants/menu.ts: CATEGORY_TO_PICK, PICKS 추가
- ErrorBoundary: interface 2개·ErrorFallback 서브컴포넌트 제거
- BoardEmpty: BoardEmptyProps·COPY 제거
- CourseRow: CourseRowProps 제거
- DayBar: DOW·헬퍼함수 4개·DayChipProps·DayChip 서브컴포넌트 제거
- MenuBoard: CATEGORY_TO_PICK·MenuBoardProps 제거
- PreMealPick: PreMealPickProps·PICKS 제거
- HeroStatus: Status 타입·nextWorkdayKey·getStatus 제거
- HomeSide: EntryInner 서브컴포넌트·ENTRY_CLASS 제거
- SiteHeader: 모듈레벨 announcements 컴포넌트 내부로 이동
