+++ src/components/board/BoardEmpty.tsx
@@ -1,25 +1,8 @@
-interface BoardEmptyProps {
-  variant: 'closed' | 'comingUp';
-  date?: string;
-  isToday?: boolean;
-  isPast?: boolean;
-}
-
-const COPY = {
-  closed: {
-    label: 'CLOSED',
-    title: '구내식당이 쉬는 날이에요',
-    body: '',
-  },
-  comingUp: {
-    label: 'COMING UP',
-    title: '다음 주 메뉴는 목요일에 공개돼요',
-    body: '영양사 선생님이 메뉴를 짜는 중이에요. 목요일에 다시 확인해 주세요',
-  },
-} as const;
+import {BOARD_EMPTY_COPY} from '@/constants/board';
+import {BoardEmptyProps} from '@/types/board';
 
 const BoardEmpty = ({variant, date, isToday, isPast}: BoardEmptyProps) => {
-  const copy = COPY[variant];
+  const copy = BOARD_EMPTY_COPY[variant];
 
   const closedTitle = (() => {
     if (variant !== 'closed') return copy.title;
