+++ src/components/board/DayBar.tsx
@@ -2,72 +2,11 @@
 
 import {useRef} from 'react';
 
-import dayjs from '@/lib/dayjs';
 import {useHasMounted} from '@/lib/useHasMounted';
 import {cn} from '@/lib/utils';
 import {useDateStore} from '@/store/useDateStore';
 
-const DOW = ['일', '월', '화', '수', '목', '금', '토'];
-
-const chipBg = (isSelected: boolean, isToday: boolean) => {
-  if (!isSelected) return 'bg-transparent hover:bg-surface-warm';
-  return isToday ? 'bg-accent' : 'bg-ink';
-};
-
-const dowColor = (dow: number) => {
-  if (dow === 0) return 'text-red-500';
-  if (dow === 6) return 'text-blue-500';
-  return null;
-};
-
-const labelClass = (isSelected: boolean, isToday: boolean, dow: number) => {
-  if (isSelected) return isToday ? 'text-ink/55' : 'text-white/60';
-  return dowColor(dow) ?? 'text-muted';
-};
-
-const dateClass = (isSelected: boolean, isToday: boolean, dow: number) => {
-  if (isSelected) return isToday ? 'text-ink' : 'text-white';
-  return dowColor(dow) ?? 'text-ink';
-};
-
-interface DayChipProps {
-  day: dayjs.Dayjs;
-  today: dayjs.Dayjs;
-  selectedDate: dayjs.Dayjs;
-  onSelect: (day: dayjs.Dayjs) => void;
-  /** 마운트 전(=SSR/정적 HTML)에는 강조를 적용하지 않아 stale '오늘' 점프를 막는다 */
-  mounted: boolean;
-}
-
-const DayChip = ({day, today, selectedDate, onSelect, mounted}: DayChipProps) => {
-  const isToday = mounted && day.isSame(today, 'day');
-  const isSelected = mounted && day.isSame(selectedDate, 'day');
-
-  return (
-    <button
-      type="button"
-      onClick={() => onSelect(day)}
-      aria-pressed={isSelected}
-      className={cn(
-        'flex min-h-[44px] flex-1 cursor-pointer flex-col items-center justify-center gap-0.5 transition-colors duration-150',
-        chipBg(isSelected, isToday)
-      )}
-    >
-      <span
-        suppressHydrationWarning
-        className={cn('text-[10px] font-bold', labelClass(isSelected, isToday, day.day()))}
-      >
-        {isToday ? '오늘' : DOW[day.day()]}
-      </span>
-      <span
-        suppressHydrationWarning
-        className={cn('text-[13.5px] font-extrabold', dateClass(isSelected, isToday, day.day()))}
-      >
-        {day.date()}
-      </span>
-    </button>
-  );
-};
+import DayChip from './DayChip';
 
 const DayBar = () => {
   const {
@@ -105,7 +44,7 @@ const DayBar = () => {
         aria-label="지난주 메뉴 보기"
         className={cn(
           'flex min-h-[44px] w-8 shrink-0 cursor-pointer flex-col items-center justify-center gap-0.5 transition-opacity duration-100 active:scale-75',
-          canGoPrev ? 'text-muted hover:text-ink' : 'cursor-default opacity-25'
+          canGoPrev ? 'text-muted hover:text-ink' : 'cursor-default opacity-25',
         )}
       >
         <span className="text-[16px] font-light leading-none">‹</span>
@@ -135,7 +74,7 @@ const DayBar = () => {
         aria-label="다음 주 메뉴 보기"
         className={cn(
           'flex min-h-[44px] w-8 shrink-0 cursor-pointer flex-col items-center justify-center gap-0.5 transition-opacity duration-100 active:scale-75',
-          canGoNext ? 'text-muted hover:text-ink' : 'cursor-default opacity-25'
+          canGoNext ? 'text-muted hover:text-ink' : 'cursor-default opacity-25',
         )}
       >
         <span className="text-[16px] font-light leading-none">›</span>
