+++ src/components/home/HomeSide.styles.ts
@@ -1,0 +1,2 @@
+export const entryClass =
+  'flex min-h-[62px] flex-1 items-center gap-3.5 bg-surface px-4 shadow-[var(--shadow-card)]';
