+++ src/constants/board.ts
@@ -1,0 +1,12 @@
+export const BOARD_EMPTY_COPY = {
+  closed: {
+    label: 'CLOSED',
+    title: '구내식당이 쉬는 날이에요',
+    body: '',
+  },
+  comingUp: {
+    label: 'COMING UP',
+    title: '다음 주 메뉴는 목요일에 공개돼요',
+    body: '영양사 선생님이 메뉴를 짜는 중이에요. 목요일에 다시 확인해 주세요',
+  },
+} as const;
