+++ src/constants/date.ts
@@ -1,0 +1,1 @@
+export const DOW = ['일', '월', '화', '수', '목', '금', '토'] as const;
