+++ src/constants/menu.ts
@@ -1,4 +1,5 @@
 import {MenuCategory} from '@/types/menu';
+import {PickType} from '@/types/vote';
 
 export const MENU_CATEGORIES: MenuCategory[] = [
   'COURSE_1',
@@ -21,3 +22,16 @@ export const MenuCategoryLabel: Record<MenuCategory, {ko: string; en: string}> =
       en: 'Take Out',
     },
   } as const;
+
+export const CATEGORY_TO_PICK: Partial<Record<MenuCategory, PickType>> = {
+  COURSE_1: 'A',
+  COURSE_2: 'B',
+  TAKE_OUT: 'takeout',
+};
+
+export const PICKS: {type: PickType; label: string}[] = [
+  {type: 'A', label: 'A코스'},
+  {type: 'B', label: 'B코스'},
+  {type: 'takeout', label: '테이크아웃'},
+  {type: 'pass', label: '패스'},
+];
