+++ src/types/shared.ts
@@ -1,0 +1,11 @@
+import type {ReactNode} from 'react';
+
+export interface ErrorBoundaryProps {
+  children: ReactNode;
+  fallback?: ReactNode;
+}
+
+export interface ErrorBoundaryState {
+  hasError: boolean;
+  error: Error | null;
+}
