+++ src/types/site.ts
@@ -1,0 +1,6 @@
+export interface EntryInnerProps {
+  no: string;
+  label: string;
+  desc: string;
+  disabled?: boolean;
+}
